Move get_utxo to the payments folder.
