+++ libsovtoken/tests/build_payment_req_handler_test.rs
@@ -280,7 +280,7 @@ pub fn build_and_submit_payment_req() {
     let Setup {addresses: payment_addresses, pool_handle, trustees, ..} = setup;
     let dids = trustees.dids();
 
-    let utxo = utils::get_utxo::get_first_utxo_txo_for_payment_address(&wallet, pool_handle, dids[0], &payment_addresses[0]);
+    let utxo = utils::payment::get_utxo::get_first_utxo_txo_for_payment_address(&wallet, pool_handle, dids[0], &payment_addresses[0]);
 
     let inputs = json!([utxo]).to_string();
     let outputs = json!([
@@ -335,7 +335,7 @@ pub fn build_and_submit_payment_req_insufficient_funds() {
     let Setup {addresses, pool_handle, trustees, ..} = setup;
     let dids = trustees.dids();
 
-    let utxo = utils::get_utxo::get_first_utxo_txo_for_payment_address(&wallet, pool_handle, dids[0], &addresses[0]);
+    let utxo = utils::payment::get_utxo::get_first_utxo_txo_for_payment_address(&wallet, pool_handle, dids[0], &addresses[0]);
 
     let inputs = json!([utxo]).to_string();
     let outputs = json!([
