+++ libsovtoken/tests/utils/payment/mod.rs
@@ -3,3 +3,4 @@ Helpers for calling payment api
 */
 
 pub mod address;
+pub mod get_utxo;
